news-website/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
openai==0.27.0
pymongo==3.12.0
schedule==1.1.0
beautifulsoup4==4.9.3
python-slugify==5.0.2
Pillow==8.3.2
